refactor: コンポーネントに移動
